:truck: rename directory's name and class name
