+++ src/main/java/com/timetoast/statistics_service/member/domain/enums/MemberRole.java
@@ -1,4 +1,4 @@
-package com.timetoast.statistics_service.member.domain.model;
+package com.timetoast.statistics_service.member.domain.enums;
 
 public enum MemberRole {
     USER, MANAGER, CREATOR, STAFF
